requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
PyWavelets>=1.4.0
pycryptodome>=3.18.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.2.0

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0
